Three-letter months in the calendar, and the watch reinstalled
TWO OF SEAN'S, and one of them was me misreading him.

"TRY AGAIN" MEANT THE WATCH, not the add button. I read it as the button still
looking too tall, took it from 26 to 22 and was measuring the label's clipping
floor when he corrected me: "no just try the watch build again, the add button
looks fine now". The 22 is reverted unshipped and the height is SETTLED at 26,
recorded in Pill so the next reader does not reopen it. The watch reinstall
itself went in on the first attempt once its tunnel came back — the failure that
prompted the ask was NWError 60, ordinary Mac↔watch flakiness.

THREE-LETTER MONTHS: "use 3 chars for month name in the calendar view like 'Aug'
instead of 'August' so the text isn't jumping around when i select days". Both
month names on the screen — the pager heading and the day panel's title — read
`month: 'short'` now, rather than a hand-rolled slice(0,3) that would be the one
date on the page disagreeing with the reader's locale.

Which of the two was actually moving is worth writing down, because the stated
reason was jumping and only one of them jumps: the pager label is CENTRED between
its arrows, so "May 2026" and "September 2026" centre at different widths and the
header shifts as you page. The day title is left-aligned and grows rightward, and
WITHIN a month its month name never changes as you select days — the width
variation there is the weekday, "Sunday" through "Wednesday". Both shortened, as
asked; the centred one is the fix.

The regression guard came for free and was verified rather than assumed:
clock.spec.ts already asserted /December 2026/ and toContainText('December 31'),
and "December 2026" does not contain the substring "Dec 2026" — so pointing them
at the short form makes them a live check, and putting the screen back to 'long'
turns both red. app.spec.ts's landing test builds the label itself and needed the
same edit. The paging specs only ever compared labels for INEQUALITY, so they
never cared which form it was — worth knowing before trusting them for this.

Web deployed to test (served index.html matches the export); build 35 on the
phone and the watch, both confirmed by asking the device rather than the
installer. Core 507, gestures 212/212, WebKit 15/15, server 48. Old derived data
deleted before the build, per the disk lesson.

+++ e2e/app.spec.ts
@@ -48,7 +48,10 @@ async function signup(page: Page): Promise<string> {
 
 test('signing up lands on the calendar, like the suite', async ({ page }) => {
   await signup(page);
-  const month = new Date().toLocaleDateString(undefined, { month: 'long', year: 'numeric' });
+  // 'short', matching the header since 2026-08-13 — Sean asked for a 3-char
+  // month so the centred label stops shifting as you page. Built the same way
+  // the screen builds it rather than hardcoded, so this holds in any locale.
+  const month = new Date().toLocaleDateString(undefined, { month: 'short', year: 'numeric' });
   await expect(page.getByText(month)).toBeVisible();
 });
